app/requirements.txt
uvicorn
fastapi
pydantic[email]
httpx

python-dotenv

beanie
fastapi-mongo-base
